Fix doc building with IPython 3
Good news and bad news. Good news is that with IPython 3, we can
actually remove a big chunk of code and just use the IPython 3 API
to export evaluated notebooks! Bad news is that, until we can rely
on developers having IPython 3, we have to keep both pieces of code
around. Yay.